requirements.txt
streamlit 
pandas 
reportlab
openpyxl 
streamlit
sentence-transformers
faiss-cpu
pypdf  